Increase width of Variant to 24 bytes to accomodate fat pointers for Gc<T> where T: ?Sized
Since the width is increasing anyways, increased the width of integers and inline strings

+++ src/language.rs
@@ -7,7 +7,7 @@ use crate::lexer::rules::keywords::KeywordRule;
 use crate::lexer::rules::literals::*;
 use crate::lexer::rules::literals::string::*;
 
-pub type IntType = i64;    // internal representation for integers
+pub type IntType = i128;    // internal representation for integers
 pub type FloatType = f64;  // internal representation for floats
 pub type InternSymbol = DefaultSymbol;  // for interned strings
